Add sync_with_odom option
